update blog roll split feature posts to use blog roll
