+++ src/components/BlogRoll/blogRoll.sass
@@ -1,0 +1,16 @@
+@import "../styles/colors"
+
+.blog-item
+  p
+    color: $sb-yellow-invert
+  &__thumbnail
+    padding: 5px
+    border: 3px solid $sb-blue
+    width: 100%
+  &__header
+    background: $sb-yellow
+    width: 80%
+    margin: 0 auto
+    transform: translate(0%, -60%)
+    padding: 15px
+
