+++ src/components/BlogRoll/index.js
@@ -1,0 +1,1 @@
+export { default } from './BlogRoll'
