+++ src/components/FeaturedPosts/featurePosts.sass
@@ -1,15 +1,6 @@
 @import "../styles/colors"
 
 .feature-post
-  &__thumbnail
-    padding: 5px
-    border: 3px solid $sb-blue
-  &__header
-    background: $sb-yellow
-    width: 80%
-    margin: 0 auto
-    transform: translate(0%, -60%)
-    padding: 15px
   &__btn.btn
     background: $sb-yellow
     color: $sb-yellow-invert
